Change: Finalization page scrolling
